Updated getting started
